requirements.txt
Flask
Pillow
requests
torch
torchvision
torchaudio
rembg